requirements.txt
fastapi==0.115.0
uvicorn==0.29.0 
click==8.1.7
SQLAlchemy==2.0.29
langchain-pinecone==0.1.0
langchainhub==0.1.15
openai==1.51.2
tiktoken==0.8.0
pypdf==4.2.0
python-dotenv==1.0.1
langchain_openai==0.1.7
langchain-core==0.1.52
langchain==0.1.20
pinecone-client==3.2.2
python-multipart==0.0.12




